added updateUser put function. Works well. Thoughts array in user do not update to new username, will fix

+++ controllers/userControllers.js
@@ -1,9 +1,9 @@
-const { User } = require('../models/models');
+const { User, Thought } = require('../models/models');
 
 // Get all users
 const getAllUsers = async (req, res) => {
   try {
-    const users = await User.find().populate('thoughts friends');
+    const users = await User.find().populate("thoughts friends");
     res.status(200).json(users);
   } catch (err) {
     res.status(500).json(err);
@@ -13,9 +13,11 @@ const getAllUsers = async (req, res) => {
 // Get a single user by ID
 const getUserById = async (req, res) => {
   try {
-    const user = await User.findById(req.params.id).populate('thoughts friends');
+    const user = await User.findById(req.params.id).populate(
+      "thoughts friends"
+    );
     if (!user) {
-      res.status(404).json({ message: 'No user found with this id!' });
+      res.status(404).json({ message: "No user found with this id!" });
       return;
     }
     res.status(200).json(user);
@@ -25,27 +27,53 @@ const getUserById = async (req, res) => {
 };
 
 const createUser = async (req, res) => {
-    try {
-      const user = await User.create(req.body);
-      res.status(200).json(user);
-    } catch (err) {
-      res.status(400).json(err);
-    }
-}
+  try {
+    const user = await User.create(req.body);
+    res.status(200).json(user);
+  } catch (err) {
+    res.status(400).json(err);
+  }
+};
 
-const deleteUser = async (req, res) => {
+const updateUser = async (req, res) => {
     try {
-        const user = await User.findByIdAndDelete(req.params.id);
+        const user = await User.findByIdAndUpdate(
+            req.params.id,
+            req.body,
+            { new: true }
+        );
         if (!user) {
-            res.status(404).json({ message: 'No user found with this id!' });
+            res.status(404).json({ message: "No user found with this id!" });
             return;
-        } else {
-            res.status(200).json({ message: `Delete successful: deleted user with id ${req.params.id}` });
         }
-          
+        res.status(200).json({ message: `Successfully updated user with id:${req.params.id}` });
     } catch (err) {
         res.status(500).json(err);
-    }  
-}
+    }
+};
 
-module.exports = { getAllUsers, getUserById,createUser, deleteUser };
+const deleteUser = async (req, res) => {
+  try {
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      res.status(404).json({ message: "No user found with this id!" });
+      return;
+    }
+
+    // Delete all of the user's thoughts
+    await Thought.deleteMany({ username: user.username });
+
+    // Delete the user
+    await User.findByIdAndDelete(req.params.id);
+
+    res
+      .status(200)
+      .json({
+        message: `Delete successful: deleted user with id:${req.params.id} and associated thoughts if applicable.`,
+      });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+};
+
+module.exports = { getAllUsers, getUserById, createUser, updateUser, deleteUser };
